requirements.txt
flask
opencv-python
numpy
matplotlib
svgwrite
svgpathtools
pillow
